Move runPentestSwarm from tool file to workflows/pentest.ts
runPentestSwarm is core orchestration logic, not tool-specific. Moving it
to the workflow module eliminates the inappropriate dependency where
workflows imported from a tool file. The spawn_pentest_swarm tool now
imports from the workflow instead.
