+++ src/core/agents/offSecAgent/tools/spawnPentestSwarm.ts
@@ -1,162 +1,11 @@
 import { tool } from "ai";
-import type { ModelMessage } from "ai";
 import { z } from "zod";
 import { FindingsRegistry } from "../../../findings/registry";
-import { TargetedPentestAgent } from "../../specialized/pentest/agent";
-import type { PentestResult } from "../../specialized/pentest/agent";
-import type { SubagentConsumeCallbacks } from "../types";
-import type { AIModel } from "../../../ai";
-import type { AIAuthConfig } from "../../../ai/utils";
-import type { SessionInfo } from "../../../session";
 import {
-  saveSubagentData,
-  writeAgentManifest,
-  buildManifestEntries,
-  finalizeManifest,
-  type SwarmTarget,
-} from "../../../session/persistence";
-import { runWithBoundedConcurrency } from "../../../utils/concurrency";
+  runPentestSwarm,
+  DEFAULT_CONCURRENCY,
+} from "../../../workflows/pentest";
 import type { ToolContext } from "./types";
-
-// ---------------------------------------------------------------------------
-// Pentest swarm orchestration
-// ---------------------------------------------------------------------------
-
-/** Default max concurrent agents */
-const DEFAULT_CONCURRENCY = 10;
-
-export interface PentestSwarmInput {
-  targets: SwarmTarget[];
-  model: AIModel;
-  session: SessionInfo;
-  authConfig?: AIAuthConfig;
-  abortSignal?: AbortSignal;
-  findingsRegistry: FindingsRegistry;
-  subagentCallbacks?: SubagentConsumeCallbacks;
-  onError?: (e: unknown) => void;
-  concurrency?: number;
-}
-
-/**
- * Run a pentest swarm: build manifest, run agents with bounded concurrency
- * and automatic persistence, then finalize the manifest.
- *
- * Each agent gets message capture, callback wrapping, spawn/complete events,
- * and persistence on success or failure — all handled internally.
- *
- * Returns `(PentestResult | null)[]` in the same order as targets.
- */
-export async function runPentestSwarm(
-  input: PentestSwarmInput,
-): Promise<(PentestResult | null)[]> {
-  const {
-    targets,
-    model,
-    session,
-    authConfig,
-    abortSignal,
-    findingsRegistry,
-    subagentCallbacks,
-    onError,
-    concurrency = DEFAULT_CONCURRENCY,
-  } = input;
-
-  // Write agent manifest with all agents as "running" — if the process
-  // crashes mid-run, the loader will detect these as paused/interrupted.
-  const manifestEntries = buildManifestEntries(targets);
-  writeAgentManifest(session, manifestEntries);
-
-  const results = await runWithBoundedConcurrency(
-    targets,
-    concurrency,
-    async (target, index) => {
-      const subagentId = `pentest-agent-${index + 1}`;
-      let lastMessages: ModelMessage[] = [];
-
-      const onStepFinish = (e: { response: { messages?: ModelMessage[] } }) => {
-        if (e.response.messages) {
-          lastMessages = e.response.messages;
-        }
-      };
-
-      const wrappedCallbacks: SubagentConsumeCallbacks | undefined =
-        subagentCallbacks
-          ? {
-              onTextDelta: (d) =>
-                subagentCallbacks.onTextDelta?.({ ...d, subagentId }),
-              onToolCall: (d) =>
-                subagentCallbacks.onToolCall?.({ ...d, subagentId }),
-              onToolResult: (d) =>
-                subagentCallbacks.onToolResult?.({ ...d, subagentId }),
-              onError: (e) => subagentCallbacks.onError?.(e),
-            }
-          : undefined;
-
-      subagentCallbacks?.onSubagentSpawn?.({
-        subagentId,
-        input: { target: target.target, objectives: target.objectives },
-        status: "pending",
-      });
-
-      try {
-        const agent = new TargetedPentestAgent({
-          target: target.target,
-          objectives: target.objectives,
-          model,
-          session,
-          authConfig,
-          abortSignal,
-          findingsRegistry,
-          onStepFinish,
-        });
-
-        const result = await agent.consume({
-          onError: (e) => onError?.(e),
-          subagentCallbacks: wrappedCallbacks,
-        });
-
-        saveSubagentData(session, {
-          agentName: subagentId,
-          target: target.target,
-          objective: target.objectives.join("; "),
-          status: "completed",
-          findingsCount: result.findings.length,
-          messages: lastMessages,
-        });
-
-        subagentCallbacks?.onSubagentComplete?.({
-          subagentId,
-          input: { target: target.target, objectives: target.objectives },
-          status: "completed",
-        });
-
-        return result;
-      } catch (error) {
-        saveSubagentData(session, {
-          agentName: subagentId,
-          target: target.target,
-          objective: target.objectives.join("; "),
-          status: "failed",
-          error: error instanceof Error ? error.message : String(error),
-          messages: lastMessages,
-        });
-
-        subagentCallbacks?.onSubagentComplete?.({
-          subagentId,
-          input: { target: target.target, objectives: target.objectives },
-          status: "failed",
-        });
-
-        throw error;
-      }
-    },
-  );
-
-  // Rewrite manifest with final statuses (index-based matching)
-  finalizeManifest(session, manifestEntries, results);
-
-  return results;
-}
 
 // ---------------------------------------------------------------------------
 // Tool factory
